Extract functionality and add placeholder for DynamoDB
