+++ internal/app/controller/apiv1/create.go
@@ -32,7 +32,7 @@ func (c *Create) Run(response http.ResponseWriter, request *http.Request) {
 			response.WriteHeader(http.StatusBadRequest)
 		} else {
 			journal := model.Journal{ID: 0, Slug: model.Slugify(journalRequest.Title), Title: journalRequest.Title, Date: journalRequest.Date, Content: journalRequest.Content}
-			js := model.Journals{Container: container, Gs: model.GiphyAdapter(container)}
+			js := model.NewJournalStore(container, model.GiphyAdapter(container))
 			journal = js.Save(journal)
 			response.WriteHeader(http.StatusCreated)
 			encoder := json.NewEncoder(response)
