+++ internal/app/controller/apiv1/list.go
@@ -17,7 +17,7 @@ type List struct {
 // Run List action
 func (c *List) Run(response http.ResponseWriter, request *http.Request) {
 
-	js := model.Journals{Container: c.Super.Container.(*app.Container), Gs: model.GiphyAdapter(c.Super.Container.(*app.Container))}
+	js := model.NewJournalStore(c.Super.Container.(*app.Container), model.GiphyAdapter(c.Super.Container.(*app.Container)))
 	journals := js.FetchAll()
 	response.Header().Add("Content-Type", "application/json")
 	encoder := json.NewEncoder(response)
