+++ internal/app/controller/apiv1/single.go
@@ -17,7 +17,7 @@ type Single struct {
 // Run Single action
 func (c *Single) Run(response http.ResponseWriter, request *http.Request) {
 
-	js := model.Journals{Container: c.Super.Container.(*app.Container), Gs: model.GiphyAdapter(c.Super.Container.(*app.Container))}
+	js := model.NewJournalStore(c.Super.Container.(*app.Container), model.GiphyAdapter(c.Super.Container.(*app.Container)))
 	journal := js.FindBySlug(c.Params[1])
 
 	response.Header().Add("Content-Type", "application/json")
