+++ journal_test.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"io/ioutil"
+	"io"
 	"log"
 	"net/http"
 	"net/http/httptest"
@@ -10,11 +10,11 @@ import (
 
 	"github.com/jamiefdhurst/journal/pkg/adapter/giphy"
 	"github.com/jamiefdhurst/journal/pkg/adapter/json"
+	"github.com/jamiefdhurst/journal/pkg/database/sqlite"
 
 	"github.com/jamiefdhurst/journal/internal/app"
 	"github.com/jamiefdhurst/journal/internal/app/model"
 	"github.com/jamiefdhurst/journal/internal/app/router"
-	"github.com/jamiefdhurst/journal/pkg/database"
 	pkgrouter "github.com/jamiefdhurst/journal/pkg/router"
 )
 
@@ -33,7 +33,7 @@ func init() {
 func fixtures(t *testing.T) {
 	container := &app.Container{Configuration: app.DefaultConfiguration()}
 	adapter := giphy.Client{Client: &json.Client{}}
-	db := &database.Sqlite{}
+	db := &sqlite.Sqlite{}
 	if err := db.Connect("test/data/test.db"); err != nil {
 		t.Error("Could not open test database for writing...")
 	}
@@ -56,7 +56,7 @@ func fixtures(t *testing.T) {
 func TestApiv1List(t *testing.T) {
 	fixtures(t)
 
-	request, err := http.NewRequest("GET", server.URL+"/api/v1/post", nil)
+	request, _ := http.NewRequest("GET", server.URL+"/api/v1/post", nil)
 
 	res, err := http.DefaultClient.Do(request)
 
@@ -69,7 +69,7 @@ func TestApiv1List(t *testing.T) {
 	}
 
 	defer res.Body.Close()
-	body, err := ioutil.ReadAll(res.Body)
+	body, _ := io.ReadAll(res.Body)
 	expected := `[{"id":3,"slug":"test-3","title":"A Final Test","date":"2018-03-01T00:00:00Z","content":"<p>Test finally!</p>"},{"id":2,"slug":"test-2","title":"Another Test","date":"2018-02-01T00:00:00Z","content":"<p>Test again!</p>"},{"id":1,"slug":"test","title":"Test","date":"2018-01-01T00:00:00Z","content":"<p>Test!</p>"}]`
 
 	// Use contains to get rid of any extra whitespace that we can discount
@@ -82,7 +82,7 @@ func TestApiv1List(t *testing.T) {
 func TestApiV1Single(t *testing.T) {
 	fixtures(t)
 
-	request, err := http.NewRequest("GET", server.URL+"/api/v1/post/test", nil)
+	request, _ := http.NewRequest("GET", server.URL+"/api/v1/post/test", nil)
 
 	res, err := http.DefaultClient.Do(request)
 
@@ -95,7 +95,7 @@ func TestApiV1Single(t *testing.T) {
 	}
 
 	defer res.Body.Close()
-	body, err := ioutil.ReadAll(res.Body)
+	body, _ := io.ReadAll(res.Body)
 	expected := `{"id":1,"slug":"test","title":"Test","date":"2018-01-01T00:00:00Z","content":"<p>Test!</p>"}`
 
 	// Use contains to get rid of any extra whitespace that we can discount
@@ -107,7 +107,7 @@ func TestApiV1Single(t *testing.T) {
 func TestApiV1Single_NotFound(t *testing.T) {
 	fixtures(t)
 
-	request, err := http.NewRequest("GET", server.URL+"/api/v1/post/random", nil)
+	request, _ := http.NewRequest("GET", server.URL+"/api/v1/post/random", nil)
 
 	res, err := http.DefaultClient.Do(request)
 
@@ -123,7 +123,7 @@ func TestApiV1Single_NotFound(t *testing.T) {
 func TestApiV1Create(t *testing.T) {
 	fixtures(t)
 
-	request, err := http.NewRequest("PUT", server.URL+"/api/v1/post", strings.NewReader(`{"title":"Test 4","date":"2018-06-01T00:00:00Z","content":"<p>Test 4!</p>"}`))
+	request, _ := http.NewRequest("PUT", server.URL+"/api/v1/post", strings.NewReader(`{"title":"Test 4","date":"2018-06-01T00:00:00Z","content":"<p>Test 4!</p>"}`))
 
 	res, err := http.DefaultClient.Do(request)
 
@@ -136,7 +136,7 @@ func TestApiV1Create(t *testing.T) {
 	}
 
 	defer res.Body.Close()
-	body, err := ioutil.ReadAll(res.Body)
+	body, _ := io.ReadAll(res.Body)
 	expected := `{"id":4,"slug":"test-4","title":"Test 4","date":"2018-06-01T00:00:00Z","content":"<p>Test 4!</p>"}`
 
 	// Use contains to get rid of any extra whitespace that we can discount
@@ -148,7 +148,7 @@ func TestApiV1Create(t *testing.T) {
 func TestApiV1Create_InvalidRequest(t *testing.T) {
 	fixtures(t)
 
-	request, err := http.NewRequest("PUT", server.URL+"/api/v1/post", nil)
+	request, _ := http.NewRequest("PUT", server.URL+"/api/v1/post", nil)
 
 	res, err := http.DefaultClient.Do(request)
 
@@ -164,7 +164,7 @@ func TestApiV1Create_InvalidRequest(t *testing.T) {
 func TestApiV1Create_MissingData(t *testing.T) {
 	fixtures(t)
 
-	request, err := http.NewRequest("PUT", server.URL+"/api/v1/post", strings.NewReader(`{"title":"Test 4"}`))
+	request, _ := http.NewRequest("PUT", server.URL+"/api/v1/post", strings.NewReader(`{"title":"Test 4"}`))
 
 	res, err := http.DefaultClient.Do(request)
 
@@ -180,7 +180,7 @@ func TestApiV1Create_MissingData(t *testing.T) {
 func TestApiV1Create_RepeatTitles(t *testing.T) {
 	fixtures(t)
 
-	request, err := http.NewRequest("PUT", server.URL+"/api/v1/post", strings.NewReader(`{"title":"Repeated","date":"2018-02-01T00:00:00Z","content":"<p>Repeated content test!</p>"}`))
+	request, _ := http.NewRequest("PUT", server.URL+"/api/v1/post", strings.NewReader(`{"title":"Repeated","date":"2018-02-01T00:00:00Z","content":"<p>Repeated content test!</p>"}`))
 	res, err := http.DefaultClient.Do(request)
 	if err != nil {
 		t.Errorf("Unexpected error: %s", err)
@@ -189,7 +189,7 @@ func TestApiV1Create_RepeatTitles(t *testing.T) {
 		t.Error("Expected 201 status code")
 	}
 
-	request, err = http.NewRequest("PUT", server.URL+"/api/v1/post", strings.NewReader(`{"title":"Repeated","date":"2019-02-01T00:00:00Z","content":"<p>Repeated content test again!</p>"}`))
+	request, _ = http.NewRequest("PUT", server.URL+"/api/v1/post", strings.NewReader(`{"title":"Repeated","date":"2019-02-01T00:00:00Z","content":"<p>Repeated content test again!</p>"}`))
 	res, err = http.DefaultClient.Do(request)
 	if err != nil {
 		t.Errorf("Unexpected error: %s", err)
@@ -198,7 +198,7 @@ func TestApiV1Create_RepeatTitles(t *testing.T) {
 		t.Error("Expected 201 status code")
 	}
 
-	request, err = http.NewRequest("GET", server.URL+"/api/v1/post/repeated-1", nil)
+	request, _ = http.NewRequest("GET", server.URL+"/api/v1/post/repeated-1", nil)
 	res, err = http.DefaultClient.Do(request)
 	if err != nil {
 		t.Errorf("Unexpected error: %s", err)
@@ -207,7 +207,7 @@ func TestApiV1Create_RepeatTitles(t *testing.T) {
 		t.Error("Expected 200 status code")
 	}
 	defer res.Body.Close()
-	body, err := ioutil.ReadAll(res.Body)
+	body, _ := io.ReadAll(res.Body)
 	expected := `{"id":5,"slug":"repeated-1","title":"Repeated","date":"2019-02-01T00:00:00Z","content":"<p>Repeated content test again!</p>"}`
 
 	// Use contains to get rid of any extra whitespace that we can discount
@@ -219,7 +219,7 @@ func TestApiV1Create_RepeatTitles(t *testing.T) {
 func TestApiV1Update(t *testing.T) {
 	fixtures(t)
 
-	request, err := http.NewRequest("POST", server.URL+"/api/v1/post/test", strings.NewReader(`{"title":"A different title"}`))
+	request, _ := http.NewRequest("POST", server.URL+"/api/v1/post/test", strings.NewReader(`{"title":"A different title"}`))
 
 	res, err := http.DefaultClient.Do(request)
 
@@ -232,7 +232,7 @@ func TestApiV1Update(t *testing.T) {
 	}
 
 	defer res.Body.Close()
-	body, err := ioutil.ReadAll(res.Body)
+	body, _ := io.ReadAll(res.Body)
 	expected := `{"id":1,"slug":"test","title":"A different title","date":"2018-01-01T00:00:00Z","content":"<p>Test!</p>"}`
 
 	// Use contains to get rid of any extra whitespace that we can discount
@@ -244,7 +244,7 @@ func TestApiV1Update(t *testing.T) {
 func TestApiV1Update_NotFound(t *testing.T) {
 	fixtures(t)
 
-	request, err := http.NewRequest("POST", server.URL+"/api/v1/post/random", strings.NewReader(`{"title":"A different title"}`))
+	request, _ := http.NewRequest("POST", server.URL+"/api/v1/post/random", strings.NewReader(`{"title":"A different title"}`))
 
 	res, err := http.DefaultClient.Do(request)
 
@@ -260,7 +260,7 @@ func TestApiV1Update_NotFound(t *testing.T) {
 func TestApiV1Update_InvalidRequest(t *testing.T) {
 	fixtures(t)
 
-	request, err := http.NewRequest("POST", server.URL+"/api/v1/post/test", nil)
+	request, _ := http.NewRequest("POST", server.URL+"/api/v1/post/test", nil)
 
 	res, err := http.DefaultClient.Do(request)
