+++ pkg/database/result/result.go
@@ -1,0 +1,7 @@
+package result
+
+// Result summarises an executed database command
+type Result interface {
+	LastInsertId() (int64, error)
+	RowsAffected() (int64, error)
+}
