+++ pkg/database/sqlite/sqlite_test.go
@@ -1,4 +1,4 @@
-package database
+package sqlite
 
 import (
 	"os"
